refactor(otel): move trace ID and span ID log valuers to dedicated package
- Extract TraceID and SpanID log valuers from middleware to kratos/log/otel package
- Add trace_context.go for trace ID and span ID valuer implementations
- Update go.mod to include go.opentelemetry.io/otel/trace as direct dependency
- Enhance README with usage example for TraceID and SpanID valuers
- Simplify tracing middleware by removing redundant log valuer functions
- Refactor tracing tests to use new traceIDs helper function
- Add unit tests for TraceID and SpanID valuer functions in trace_context_test.go
