+++ kratos/log/otel/trace_context.go
@@ -1,0 +1,28 @@
+package otel
+
+import (
+	"context"
+
+	"github.com/go-kratos/kratos/v2/log"
+	"go.opentelemetry.io/otel/trace"
+)
+
+// TraceID returns a log valuer for the current span trace ID.
+func TraceID() log.Valuer {
+	return func(ctx context.Context) any {
+		if span := trace.SpanContextFromContext(ctx); span.HasTraceID() {
+			return span.TraceID().String()
+		}
+		return ""
+	}
+}
+
+// SpanID returns a log valuer for the current span ID.
+func SpanID() log.Valuer {
+	return func(ctx context.Context) any {
+		if span := trace.SpanContextFromContext(ctx); span.HasSpanID() {
+			return span.SpanID().String()
+		}
+		return ""
+	}
+}
