+++ kratos/middleware/otel/tracing/tracing.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/go-fries/fries/kratos/middleware/otel/tracing/v3/internal/semconv"
-	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware"
 	"github.com/go-kratos/kratos/v2/transport"
 	"go.opentelemetry.io/otel/trace"
@@ -41,25 +40,5 @@ func Client(opts ...Option) middleware.Middleware {
 			}
 			return handler(ctx, req)
 		}
-	}
-}
-
-// TraceID returns a log valuer for the current span trace ID.
-func TraceID() log.Valuer {
-	return func(ctx context.Context) any {
-		if span := trace.SpanContextFromContext(ctx); span.HasTraceID() {
-			return span.TraceID().String()
-		}
-		return ""
-	}
-}
-
-// SpanID returns a log valuer for the current span ID.
-func SpanID() log.Valuer {
-	return func(ctx context.Context) any {
-		if span := trace.SpanContextFromContext(ctx); span.HasSpanID() {
-			return span.SpanID().String()
-		}
-		return ""
 	}
 }
